+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gouvfr-lasuite/ui-kit
 
+[UNRELEASED]
+
+### Patch Changes
+
+- 🐛(dropdown) fix to hide icon when no icon prop is passed
+
 ## 0.20.1
 
 ### Patch Changes